introduce seeds for random() and noise()
So the user can get the exact same image by copying the URL that
contains a seed. If a seed is given in the URL, that is used. If not,
one is generated. The seed is shown on the page.

There is still a bug: If you press Ctrl-R, a new seed is generated even
though the URL contains a seed.